+++ src/main/java/org/example/jspcafe/question/controller/QuestionServlet.java
@@ -17,11 +17,14 @@ import java.util.List;
 @WebServlet("/questions")
 public class QuestionServlet extends HttpServlet {
     private QuestionRepository questionRepository;
+    private QuestionService questionService;
 
     @Override
     public void init(ServletConfig config) {
         ServletContext context = config.getServletContext();
         this.questionRepository = (QuestionRepository) context.getAttribute("QuestionRepository");
+        this.questionService = (QuestionService) context.getAttribute("QuestionService");
+
     }
 
     @Override
@@ -29,5 +32,15 @@ public class QuestionServlet extends HttpServlet {
         List<Question> all = questionRepository.getAll();
         req.setAttribute("questions", all);
         req.getRequestDispatcher("/qna/list.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        String writer = req.getParameter("writer");
+        String title = req.getParameter("title");
+        String contents = req.getParameter("contents");
+        Question question = new Question(writer, title, contents);
+        Long id = questionService.saveQuestion(question);
+        resp.sendRedirect("/questions/" + id);
     }
 }
